add the github api hooks to the seed file and set up a hook library to manage hooks on repositories
